+++ src/components/connected/statistics/BottomPage.js
@@ -1,10 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { useDispatch, useSelector } from 'react-redux';
-import moment from 'moment';
 import 'moment/locale/fr';
 import ElementHighcharts from './Components/ElementHighcharts';
-import { sortByMonthAndYear } from '../../../utils/functionsSort';
 
 import labelsCorrespondance from '../../../data/labelsCorrespondance.json';
 import { statistiqueActions } from '../../../actions/statistique.actions';
@@ -18,55 +16,11 @@ function BottomPage({ donneesStats, print, type }) {
   const tabColorAge = ['#ff007a', '#6945bd', '#c6c9ae', '#ff5e3b', '#00ba8e'];
   const tabColorStatut = ['#a2b4b1', '#ffdbd2', '#a3a6bc', '#ddb094', '#fff480'];
   const tabColorLieux = ['#ff007a', '#6945bd', '#c6c9ae', '#ff5e3b', '#00ba8e', '#a2b4b1', '#ffdbd2', '#a3a6bc', '#ddb094', '#fff480'];
+  const tabColorTemps = ['', '#76adf8', '#a96751', '#496b30'];
+  const tabColorDuree = ['#abcdf5', '#abcdf5', '#abcdf5', '#abcdf5'];
+  const { statsUsagers, statsAges, statsTempsAccompagnements, statsDurees, statsLieux, statsReorientations } = donneesStats;
 
-  const get4lastMonths = (month, year) => {
-    let monthToPrint = [month];
-    let yearAssociated = [year];
-    let lastInsertedMonth = month;
-    let lastInsertedYear = year;
-    for (let i = 0; i < 3; i++) {
-      lastInsertedYear = lastInsertedMonth - 1 === -1 ? lastInsertedYear - 1 : lastInsertedYear;
-      lastInsertedMonth = lastInsertedMonth - 1 === -1 ? 11 : lastInsertedMonth - 1; //11 = décembre dans Date
-      monthToPrint.push(lastInsertedMonth);
-      yearAssociated.push(lastInsertedYear.toString());
-    }
-    return [monthToPrint, yearAssociated];
-  };
 
-  const { statsEvolutions, statsUsagers, statsAges, statsReorientations } = donneesStats;
-
-  //Map des stats evolutions pour ajouter les données nécessaires pour le graph (label mois année, valeur)
-  let statsEvolutionsMapped = [];
-  for (const [annee, moisListe] of Object.entries(statsEvolutions)) {
-    let statsEvolutionsMapped2 = moisListe.map(mois => {
-      mois.nom = moment().month(`${mois.mois}`).format('MMMM');
-      mois.nom = mois.nom?.concat(' ', annee);
-      mois.annee = annee;
-      mois.valeur = mois.totalCras;
-      return mois;
-    });
-    statsEvolutionsMapped.push(...statsEvolutionsMapped2);
-  }
-
-  //Filtrage pour ne garder que le mois en cours et les 3 précédents max
-  let monthToPrint = get4lastMonths(new Date().getMonth(), new Date().getUTCFullYear());
-  let statsEvolutionsFiltered = Object.values(statsEvolutionsMapped).filter(mois => {
-    // eslint-disable-next-line max-len
-    return monthToPrint[0].includes(mois.mois) && monthToPrint[1][monthToPrint[0].findIndex(mois2 => mois.mois === mois2)].toString() === mois.annee ? mois : '';
-  });
-
-  //Ajout des mois manquants (donc avec totalCras à 0)
-  monthToPrint[0].forEach((value, index) => {
-    if (statsEvolutionsFiltered.some(mois => mois.mois === value) === false) {
-      let annee = monthToPrint[1][index];
-      let nom = moment().month(`${value}`).format('MMMM');
-      nom = nom?.concat(' ', annee);
-      statsEvolutionsFiltered.push({ 'mois': value, 'valeur': 0, 'annee': annee.toString(), 'nom': nom });
-    }
-  });
-
-  //Tri par mois/annee croissant
-  statsEvolutionsFiltered.sort(sortByMonthAndYear);
 
   //Tri liste des réorientations autres
   if (statsReorientations?.length > 0) {
@@ -98,43 +52,6 @@ function BottomPage({ donneesStats, print, type }) {
       dispatch(statistiqueActions.updateListeAutresReorientations(listeAutresReorientations));
     }
   }
-
-  const graphiqueEvolution = {
-    graphique: {
-      typeGraphique: 'xy',
-      largeurGraphique: 320,
-      hauteurGraphique: 310,
-      margeGaucheGraphique: 80,
-      margeDroiteGraphique: 80,
-      optionResponsive: false,
-      couleursGraphique: tabColorAge,
-      largeurGraphiquePrint: 1100,
-      hauteurGraphiquePrint: 500,
-      margeGaucheGraphiquePrint: 100,
-      margeDroiteGraphiquePrint: 200,
-    },
-    titre: {
-      optionTitre: '&Eacute;volution des comptes rendus d&rsquo;activit&eacute;',
-      margeTitre: 48,
-    }
-  };
-
-  const graphiqueEvolutionSM = {
-    graphique: {
-      typeGraphique: 'xy',
-      largeurGraphique: 320,
-      hauteurGraphique: 310,
-      margeGaucheGraphique: 40,
-      margeDroiteGraphique: 70,
-      optionResponsive: false,
-      couleursGraphique: tabColorAge
-    },
-    titre: {
-      optionTitre: '&Eacute;volution des comptes rendus d&rsquo;activit&eacute',
-      margeTitre: 48,
-      placementTitre: 10,
-    }
-  };
 
   const graphiqueAge = {
     graphique: {
@@ -173,6 +90,120 @@ function BottomPage({ donneesStats, print, type }) {
     titre: {
       optionTitre: 'Statut des usagers',
       margeTitre: 34,
+    }
+  };
+
+  const pieGraphique = {
+    graphique: {
+      typeGraphique: 'pie',
+      largeurGraphique: 300,
+      hauteurGraphique: 320,
+      margeGaucheGraphique: 0,
+      margeDroiteGraphique: 10,
+      optionResponsive: false,
+      couleursGraphique: tabColorLieux,
+      largeurGraphiquePrint: 1100,
+      hauteurGraphiquePrint: 400,
+      margeGaucheGraphiquePrint: 0,
+      margeDroiteGraphiquePrint: 10,
+    },
+    titre: {
+      optionTitre: 'Canaux d&rsquo;accompagnements',
+      margeTitre: 48,
+      placementTitre: 0
+    }
+  };
+
+  const pieGraphiqueSm = {
+    graphique: {
+      typeGraphique: 'pie',
+      largeurGraphique: 300,
+      hauteurGraphique: 320,
+      margeGaucheGraphique: 0,
+      margeDroiteGraphique: 10,
+      optionResponsive: true,
+      couleursGraphique: tabColorLieux
+    },
+    titre: {
+      optionTitre: 'Canaux d&rsquo;accompagnements',
+      margeTitre: -17,
+      placementTitre: 0
+    }
+  };
+
+  const pieGraphiqueTemps = {
+    graphique: {
+      typeGraphique: 'pie',
+      largeurGraphique: 300,
+      hauteurGraphique: 320,
+      margeGaucheGraphique: 0,
+      margeDroiteGraphique: 10,
+      optionResponsive: false,
+      couleursGraphique: tabColorTemps,
+      largeurGraphiquePrint: 1100,
+      hauteurGraphiquePrint: 400,
+      margeGaucheGraphiquePrint: 0,
+      margeDroiteGraphiquePrint: 10,
+    },
+    titre: {
+      optionTitre: 'Temps en accompagnement',
+      margeTitre: 48,
+      placementTitre: 0
+    }
+  };
+
+  const pieGraphiqueSmTemps = {
+    graphique: {
+      typeGraphique: 'pie',
+      largeurGraphique: 300,
+      hauteurGraphique: 320,
+      margeGaucheGraphique: 0,
+      margeDroiteGraphique: 10,
+      optionResponsive: true,
+      couleursGraphique: tabColorTemps
+    },
+    titre: {
+      optionTitre: 'Temps en accompagnement',
+      margeTitre: -17,
+      placementTitre: 0
+    }
+  };
+
+  const columnGraphique = {
+    graphique: {
+      typeGraphique: 'column',
+      largeurGraphique: 360,
+      hauteurGraphique: 310,
+      margeGaucheGraphique: 55,
+      margeDroiteGraphique: 55,
+      optionResponsive: false,
+      couleursGraphique: tabColorDuree,
+      largeurGraphiquePrint: 1100,
+      hauteurGraphiquePrint: 450,
+      margeGaucheGraphiquePrint: 55,
+      margeDroiteGraphiquePrint: 55,
+    },
+    titre: {
+      optionTitre: 'Dur&eacute;e des accompagnements',
+      margeTitre: 48,
+      placementTitre: 0
+    }
+  };
+
+  const columnGraphiqueSm = {
+    graphique: {
+      typeGraphique: 'column',
+      largeurGraphique: 300,
+      hauteurGraphique: 310,
+      margeGaucheGraphique: 55,
+      margeDroiteGraphique: 55,
+      optionResponsive: false,
+      couleursGraphique: tabColorDuree
+    },
+    titre: {
+      optionTitre: 'Dur&eacute;e des accompagnements',
+      margeTitre: 48,
+      placementTitre: 0
     }
   };
 
@@ -217,25 +248,12 @@ function BottomPage({ donneesStats, print, type }) {
     <>
       <div className="fr-col-12 dont-print">
         <div className="fr-grid-row">
-
-          <div className="fr-col-12 fr-col-md-5 fr-col-lg-3 evolution-print">
-            <div className="fr-mt-6w fr-mb-5w fr-m-xs-to-md-7v dont-print"><hr/></div>
-            <span className="graphique-responsive-md-lg ">
-              <ElementHighcharts donneesStats={statsEvolutionsFiltered} variablesGraphique={graphiqueEvolution} print={print}/>
-            </span>
-            <span className="graphique-responsive-sm">
-              <ElementHighcharts donneesStats={statsEvolutionsFiltered} variablesGraphique={graphiqueEvolutionSM} print={print}/>
-            </span>
-          </div>
-
-          <div className="fr-col-offset-12 fr-col-offset-md-1"></div>
-
-          <div className="fr-col-12 fr-col-md-5 fr-col-lg-3 age-print">
+          <div className="fr-col-12 fr-col-md-5 fr-col-lg-6 age-print">
             <div className="fr-mt-6w fr-mb-5w fr-m-xs-to-md-7v"><hr className="dont-print"/></div>
             <ElementHighcharts donneesStats={statsAges} variablesGraphique={graphiqueAge} print={print}/>
           </div>
-
-          <div className="fr-col-12 fr-col-md-5 graphique-responsive-md dont-print">
+          {/**
+           <div className="fr-col-12 fr-col-md-5 graphique-responsive-md dont-print">
             {statsReorientations?.length > 0 &&
               <>
                 <div className="fr-mt-6w fr-mb-5w fr-m-xs-to-md-7v"><hr/></div>
@@ -243,11 +261,26 @@ function BottomPage({ donneesStats, print, type }) {
               </>
             }
           </div>
-
-          <div className="fr-col-offset-md-1 fr-col-12 fr-col-md-5 fr-col-lg-3 statut-print">
+          */}
+          <div className="fr-col-12 fr-col-md-5 fr-col-lg-6 statut-print">
             <div className="fr-mt-6w fr-mb-5w fr-m-xs-to-md-7v dont-print"><hr/></div>
             <ElementHighcharts donneesStats={statsUsagers} variablesGraphique={graphiqueStatut} print={print}/>
           </div>
+
+          <div className="fr-col-12 fr-col-md-4 dont-print">
+            <ElementHighcharts donneesStats={statsLieux} variablesGraphique={pieGraphique} print={print}/>
+          </div>
+
+          <div className="fr-col-12 fr-col-md-4 dont-print">
+            <ElementHighcharts donneesStats={statsTempsAccompagnements} variablesGraphique={pieGraphiqueTemps} print={print}/>
+          </div>
+
+          <div className="fr-col-12 fr-col-lg-4 dont-print">
+            <ElementHighcharts donneesStats={statsDurees} variablesGraphique={columnGraphiqueSm} print={print}/>
+          </div>
+
+
+
           <div className={type === 'conseiller' ? 'fr-col-12 fr-col-offset-md-4 fr-col-md-8 graphique-responsive-lg reorientation-print-conseiller' :
             'fr-col-12 fr-col-offset-md-4 fr-col-md-8 graphique-responsive-lg reorientation-print'}
           >
@@ -258,6 +291,11 @@ function BottomPage({ donneesStats, print, type }) {
             }
             <div className="fr-m-no-reorientation"></div>
           </div>
+        </div>
+      </div>
+
+
+      {/*
         </div>
       </div>
       <div className="fr-col-12 fr-mb-6w only-print" >
@@ -275,6 +313,8 @@ function BottomPage({ donneesStats, print, type }) {
             listeAutres={listeAutresReorientations} print={true}/>
         }
       </div>
+
+    */}
     </>
   );
 }
