+++ src/components/connected/statistics/RightPage.js
@@ -1,14 +1,100 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import moment from 'moment';
+import { sortByMonthAndYear } from '../../../utils/functionsSort';
 import ElementHighcharts from './Components/ElementHighcharts';
 
 function RightPage({ donneesStats, print }) {
 
   const tabColorTheme = ['#cac5b0', '#abb8df', '#fdcf41', '#169b62', '#80d5c6', '#ff8d7e', '#714753', '#956052', '#ddb094', '#5770be', '#ffed33', '#be9b31'];
-  const tabColorLieux = ['#ffcc9f', '#ff8d7e', '#466964', '#5770be'];
-  const tabColorDuree = ['#abcdf5', '#abcdf5', '#abcdf5', '#abcdf5'];
+  const tabColorAge = ['#ff007a', '#6945bd', '#c6c9ae', '#ff5e3b', '#00ba8e'];
 
-  const { statsThemes, statsDurees, statsLieux } = donneesStats;
+  const get4lastMonths = (month, year) => {
+    let monthToPrint = [month];
+    let yearAssociated = [year];
+    let lastInsertedMonth = month;
+    let lastInsertedYear = year;
+    for (let i = 0; i < 3; i++) {
+      lastInsertedYear = lastInsertedMonth - 1 === -1 ? lastInsertedYear - 1 : lastInsertedYear;
+      lastInsertedMonth = lastInsertedMonth - 1 === -1 ? 11 : lastInsertedMonth - 1; //11 = décembre dans Date
+      monthToPrint.push(lastInsertedMonth);
+      yearAssociated.push(lastInsertedYear.toString());
+    }
+    return [monthToPrint, yearAssociated];
+  };
+
+  const { statsThemes, statsEvolutions } = donneesStats;
+
+
+  //Map des stats evolutions pour ajouter les données nécessaires pour le graph (label mois année, valeur)
+  let statsEvolutionsMapped = [];
+  for (const [annee, moisListe] of Object.entries(statsEvolutions)) {
+    let statsEvolutionsMapped2 = moisListe.map(mois => {
+      mois.nom = moment().month(`${mois.mois}`).format('MMMM');
+      mois.nom = mois.nom?.concat(' ', annee);
+      mois.annee = annee;
+      mois.valeur = mois.totalCras;
+      return mois;
+    });
+    statsEvolutionsMapped.push(...statsEvolutionsMapped2);
+  }
+
+  //Filtrage pour ne garder que le mois en cours et les 3 précédents max
+  let monthToPrint = get4lastMonths(new Date().getMonth(), new Date().getUTCFullYear());
+  let statsEvolutionsFiltered = Object.values(statsEvolutionsMapped).filter(mois => {
+    // eslint-disable-next-line max-len
+    return monthToPrint[0].includes(mois.mois) && monthToPrint[1][monthToPrint[0].findIndex(mois2 => mois.mois === mois2)].toString() === mois.annee ? mois : '';
+  });
+
+  //Ajout des mois manquants (donc avec totalCras à 0)
+  monthToPrint[0].forEach((value, index) => {
+    if (statsEvolutionsFiltered.some(mois => mois.mois === value) === false) {
+      let annee = monthToPrint[1][index];
+      let nom = moment().month(`${value}`).format('MMMM');
+      nom = nom?.concat(' ', annee);
+      statsEvolutionsFiltered.push({ 'mois': value, 'valeur': 0, 'annee': annee.toString(), 'nom': nom });
+    }
+  });
+
+  //Tri par mois/annee croissant
+  statsEvolutionsFiltered.sort(sortByMonthAndYear);
+
+  const graphiqueEvolution = {
+    graphique: {
+      typeGraphique: 'xy',
+      largeurGraphique: 320,
+      hauteurGraphique: 310,
+      margeGaucheGraphique: 80,
+      margeDroiteGraphique: 80,
+      optionResponsive: false,
+      couleursGraphique: tabColorAge,
+      largeurGraphiquePrint: 1100,
+      hauteurGraphiquePrint: 500,
+      margeGaucheGraphiquePrint: 100,
+      margeDroiteGraphiquePrint: 200,
+    },
+    titre: {
+      optionTitre: '&Eacute;volution des comptes rendus d&rsquo;activit&eacute;',
+      margeTitre: 48,
+    }
+  };
+
+  const graphiqueEvolutionSM = {
+    graphique: {
+      typeGraphique: 'xy',
+      largeurGraphique: 320,
+      hauteurGraphique: 310,
+      margeGaucheGraphique: 40,
+      margeDroiteGraphique: 70,
+      optionResponsive: false,
+      couleursGraphique: tabColorAge
+    },
+    titre: {
+      optionTitre: '&Eacute;volution des comptes rendus d&rsquo;activit&eacute',
+      margeTitre: 48,
+      placementTitre: 10,
+    }
+  };
 
   const barGraphique = {
     graphique: {
@@ -48,87 +134,12 @@ function RightPage({ donneesStats, print }) {
     }
   };
 
-  const pieGraphique = {
-    graphique: {
-      typeGraphique: 'pie',
-      largeurGraphique: 300,
-      hauteurGraphique: 320,
-      margeGaucheGraphique: 0,
-      margeDroiteGraphique: 10,
-      optionResponsive: false,
-      couleursGraphique: tabColorLieux,
-      largeurGraphiquePrint: 1100,
-      hauteurGraphiquePrint: 400,
-      margeGaucheGraphiquePrint: 0,
-      margeDroiteGraphiquePrint: 10,
-    },
-    titre: {
-      optionTitre: 'Canaux d&rsquo;accompagnements',
-      margeTitre: 48,
-      placementTitre: 0
-    }
-  };
-
-  const pieGraphiqueSm = {
-    graphique: {
-      typeGraphique: 'pie',
-      largeurGraphique: 300,
-      hauteurGraphique: 320,
-      margeGaucheGraphique: 0,
-      margeDroiteGraphique: 10,
-      optionResponsive: true,
-      couleursGraphique: tabColorLieux
-    },
-    titre: {
-      optionTitre: 'Canaux d&rsquo;accompagnements',
-      margeTitre: -17,
-      placementTitre: 0
-    }
-  };
-
-  const columnGraphique = {
-    graphique: {
-      typeGraphique: 'column',
-      largeurGraphique: 360,
-      hauteurGraphique: 310,
-      margeGaucheGraphique: 55,
-      margeDroiteGraphique: 55,
-      optionResponsive: false,
-      couleursGraphique: tabColorDuree,
-      largeurGraphiquePrint: 1100,
-      hauteurGraphiquePrint: 450,
-      margeGaucheGraphiquePrint: 55,
-      margeDroiteGraphiquePrint: 55,
-    },
-    titre: {
-      optionTitre: 'Dur&eacute;e des accompagnements',
-      margeTitre: 48,
-      placementTitre: 0
-    }
-  };
-
-  const columnGraphiqueSm = {
-    graphique: {
-      typeGraphique: 'column',
-      largeurGraphique: 300,
-      hauteurGraphique: 310,
-      margeGaucheGraphique: 55,
-      margeDroiteGraphique: 55,
-      optionResponsive: false,
-      couleursGraphique: tabColorDuree
-    },
-    titre: {
-      optionTitre: 'Dur&eacute;e des accompagnements',
-      margeTitre: 48,
-      placementTitre: 0
-    }
-  };
-
   return (
     <>
       <div className="fr-col-12 fr-col-md-5 fr-col-lg-7 graphique-responsive-lg dont-print">
         <div className="fr-container-fluid">
           <div className="fr-grid-row ">
+
             <div className="fr-col-12">
               <ElementHighcharts donneesStats={statsThemes} variablesGraphique={barGraphique} print={print}/>
             </div>
@@ -136,9 +147,15 @@ function RightPage({ donneesStats, print }) {
             <div className="fr-col-12">
               <div className="fr-my-6w fr-m-xs-to-md-7v"><hr/></div>
             </div>
+            <div className="fr-col-12 evolution-print">
+              <span className="graphique-responsive-md-lg ">
+                <ElementHighcharts donneesStats={statsEvolutionsFiltered} variablesGraphique={graphiqueEvolution} print={print}/>
+              </span>
+            </div>
 
+            {/*
             <div className="fr-col-12 fr-col-lg-6">
-              <ElementHighcharts donneesStats={statsLieux} variablesGraphique={pieGraphiqueSm} print={print}/>
+              <ElementHighcharts donneesStats={statsTempsAccompagnements} variablesGraphique={pieGraphiqueSmTemps} print={print}/>
             </div>
 
             <div className="fr-col-12 fr-col-lg-6 dont-print">
@@ -146,6 +163,7 @@ function RightPage({ donneesStats, print }) {
                 <ElementHighcharts donneesStats={statsDurees} variablesGraphique={columnGraphiqueSm} print={print}/>
               </div>
             </div>
+           */}
           </div>
         </div>
       </div>
@@ -154,9 +172,11 @@ function RightPage({ donneesStats, print }) {
 
         <div className="fr-container-fluid">
           <div className="fr-grid-row ">
+            {/*
             <div className="fr-col-12">
               <ElementHighcharts donneesStats={statsThemes} variablesGraphique={barGraphiqueSm} print={print}/>
             </div>
+            */}
           </div>
         </div>
       </div>
@@ -164,8 +184,7 @@ function RightPage({ donneesStats, print }) {
       <div className="fr-col-12 graphique-responsive-md dont-print">
         <div className="fr-container-fluid">
           <div className="fr-grid-row ">
-
-            <div className="fr-col-12 fr-col-md-5 hr-sm-hide">
+            {/*  <div className="fr-col-12 fr-col-md-5 hr-sm-hide">
               <div className="fr-m-6w fr-m-xs-to-md-7v"><hr/></div>
             </div>
 
@@ -175,9 +194,6 @@ function RightPage({ donneesStats, print }) {
               <div className="fr-m-xs-to-md-7v"><hr/></div>
             </div>
 
-            <div className="fr-col-12 fr-col-md-6 dont-print">
-              <ElementHighcharts donneesStats={statsLieux} variablesGraphique={pieGraphiqueSm} print={print}/>
-            </div>
 
             <div className="fr-col-12 hr-md-hide dont-print">
               <div className="fr-m-6w fr-m-xs-to-md-7v"><hr/></div>
@@ -185,19 +201,23 @@ function RightPage({ donneesStats, print }) {
 
             <div className="fr-col-12 fr-col-md-6">
               <ElementHighcharts donneesStats={statsDurees} variablesGraphique={columnGraphiqueSm} print={print}/>
-            </div>
+            </div>*/}
+
           </div>
         </div>
       </div>
       <div className="fr-col-12 fr-mb-6w only-print theme-print">
-        <div className="mozilla-espace-block"></div>
-        <ElementHighcharts donneesStats={statsThemes} variablesGraphique={barGraphique} print={true}/>
-      </div>
-      <div className="fr-col-12 only-print lieux-print">
-        <ElementHighcharts donneesStats={statsLieux} variablesGraphique={pieGraphique} print={true}/>
-      </div>
-      <div className="fr-col-12 fr-mb-6w only-print duree-print">
-        <ElementHighcharts donneesStats={statsDurees} variablesGraphique={columnGraphique} print={true}/>
+        {/*
+          <div className="mozilla-espace-block"></div>
+           <ElementHighcharts donneesStats={statsThemes} variablesGraphique={barGraphique} print={true}/>
+          </div>
+          <div className="fr-col-12 only-print lieux-print">
+            <ElementHighcharts donneesStats={statsTempsAccompagnements} variablesGraphique={pieGraphiqueTemps} print={true}/>
+          </div>
+          <div className="fr-col-12 fr-mb-6w only-print duree-print">
+            <ElementHighcharts donneesStats={statsDurees} variablesGraphique={columnGraphique} print={true}/>
+          </div>
+         */}
       </div>
     </>
   );
